[MISC] Move alignment algorithm execution into core.
