+++ include/seqan3/core/algorithm/detail/all.hpp
@@ -5,12 +5,19 @@
 // shipped with this file and also available at: https://github.com/seqan/seqan3/blob/master/LICENSE.md
 // -----------------------------------------------------------------------------------------------------
 
-#include <gtest/gtest.h>
+/*!\file
+ * \brief Meta-header for the \link execution pairwise execution submodule \endlink.
+ * \author Rene Rahn <rene.rahn AT fu-berlin.de>
+ */
 
-#include <seqan3/alignment/pairwise/execution/execution_handler_parallel.hpp>
+#pragma once
 
-#include "execution_handler_template.hpp"
+#include <seqan3/core/algorithm/detail/alignment_executor_two_way.hpp>
+#include <seqan3/core/algorithm/detail/execution_handler_sequential.hpp>
 
-INSTANTIATE_TYPED_TEST_SUITE_P(execution_handler_parallel,
-                               execution_handler,
-                               seqan3::detail::execution_handler_parallel, );
+//!\cond DEV
+/*!\defgroup execution Execution
+ * \ingroup pairwise_alignment
+ * \brief Provides execution handler and executors for pairwise alignments.
+ */
+//!\endcond
