+++ test/unit/core/algorithm/CMakeLists.txt
@@ -1,3 +1,5 @@
 seqan3_test (bound_test.cpp)
 seqan3_test (configuration_test.cpp)
 seqan3_test (pipeable_config_element_test.cpp)
+
+add_subdirectories()
